Remove SovrynPriceFetcherHelper
I needed to removed this intermidiate class since it was not possible to
do proper dependency injection with this class inside
  SovrynPriceFetcher.
